+++ tests/integration/test_bayesian_optimization.py
@@ -17,7 +17,7 @@ import pytest
 import tensorflow as tf
 
 from tests.util.misc import random_seed
-from trieste.acquisition.function import BatchMonteCarloExpectedImprovement,  ExpectedHypervolumeImprovement
+from trieste.acquisition.function import BatchMonteCarloExpectedImprovement
 from trieste.acquisition.rule import (
     OBJECTIVE,
     AcquisitionRule,
@@ -28,10 +28,9 @@ from trieste.acquisition.rule import (
 from trieste.bayesian_optimizer import BayesianOptimizer
 from trieste.data import Dataset
 from trieste.models import GaussianProcessRegression
-from trieste.models.model_interfaces import ModelStack
 from trieste.space import Box
 from trieste.utils.objectives import BRANIN_MINIMIZERS, BRANIN_MINIMUM, branin, mk_observer
-from trieste.utils.multi_objectives import VLMOP2
+
 
 @random_seed
 @pytest.mark.parametrize(
@@ -82,56 +81,3 @@ def test_optimizer_finds_minima_of_the_branin_function(
     # this is a regression test
     assert tf.reduce_any(tf.reduce_all(relative_minimizer_err < 0.03, axis=-1), axis=0)
     npt.assert_allclose(best_y, BRANIN_MINIMUM, rtol=0.03)
-
-
-
-
-@random_seed
-@pytest.mark.parametrize(
-    "num_steps, acquisition_rule",
-    [
-        (20,  EfficientGlobalOptimization(ExpectedHypervolumeImprovement().using(OBJECTIVE))),
-    ],
-)
-def test_multi_objective_optimizer_finds_pareto_front_of_the_VLMOP2_function(
-    num_steps: int, acquisition_rule: AcquisitionRule
-) -> None:
-    search_space = Box([-2, -2], [2, 2])
-
-    def build_stacked_independent_objectives_model(data: Dataset) -> GaussianProcessRegression:
-        gprs =[]
-        for idx in range(2):
-            single_obj_data = Dataset(data.query_points, tf.gather(data.observations, [idx], axis=1))
-            variance = tf.math.reduce_variance(single_obj_data.observations)
-            kernel = gpflow.kernels.Matern52(variance, tf.constant([0.2, 0.2], tf.float64))
-            gpr = gpflow.models.GPR((single_obj_data.query_points, single_obj_data.observations), kernel, noise_variance=1e-5)
-            gpflow.utilities.set_trainable(gpr.likelihood, False)
-            gprs.append(GaussianProcessRegression(gpr))
-
-        return ModelStack(*gprs)
-
-
-    vlmop2 = VLMOP2().objective()
-    observer = trieste.utils.objectives.mk_observer(vlmop2, OBJECTIVE)
-
-    initial_query_points = search_space.sample(10)
-    initial_data = observer(initial_query_points)
-    model = build_stacked_independent_objectives_model(initial_data[OBJECTIVE])
-
-
-    dataset = (
-        BayesianOptimizer(observer, search_space)
-        .optimize(num_steps, initial_data, {OBJECTIVE: model}, acquisition_rule)
-        .try_get_final_datasets()[OBJECTIVE]
-    )
-
-
-
-    obs_hv = Pareto(observations).hypervolume_indicator(ref_point)
-    
-    ideal_pf = VLMOP2().gen_pareto_optimal_points(100)
-    ref_point = get_reference_point(data_observations)
-    idea_hv = Pareto(tf.cast(ideal_pf, dtype=data_observations.dtype)).hypervolume_indicator(ref_point)
-
-    log_hv_diff = math.log(idea_hv - obs_hv)
-
